moved progress bar to own component
